2025/efficient-python-dockerfile/leadspotr/pyproject.toml
[project]
name = "leadspotr"
version = "0.1.0"
description = "The Backend API for LeadSpotr service"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.95.0",
    "uvicorn>=0.21.1",
    "sqlalchemy>=2.0.7",
    "psycopg2-binary>=2.9.5",
    "sendgrid>=6.10.0",
    "python-jose[cryptography]>=3.3.0",
    "pydantic[dotenv]>=1.10.7",
    "pg8000>=1.29.4",
    "cloud-sql-python-connector>=1.2.2",
    "python-dotenv>=1.0.0",
    "google-cloud-storage>=2.9.0",
    "python-multipart>=0.0.6",
    "weasyprint>=58.1",
    "pyhumps>=3.8.0",
    "alembic>=1.11.1",
    "passlib>=1.7.4",
    "bcrypt>=4.0.1"
]
